comment

+++ assystor-api/database/migrations/2025_04_30_104520_add_status_comment_employee_to_customer_customer_group.php
@@ -9,17 +9,19 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
+public function up(): void
     {
-        Schema::table('customer_customer_group', function (Blueprint $table) {
-            $table->dropColumn('comment'); // حذف فقط
-        });
-    }
-    
-    public function down(): void
+        if (Schema::hasColumn('customer_customer_group', 'comment')) {
+            Schema::table('customer_customer_group', function (Blueprint $table) {
+                $table->dropColumn('comment');
+            });
+        }
+    }    
+public function down(): void
     {
-        Schema::table('customer_customer_group', function (Blueprint $table) {
-            $table->text('comment')->nullable(); // استعادة الحقل عند rollback
-        });
-    }
-    };
+        if (!Schema::hasColumn('customer_customer_group', 'comment')) {
+            Schema::table('customer_customer_group', function (Blueprint $table) {
+                $table->text('comment')->nullable();
+            });
+        }
+    }    };
